Add files via upload
Adding in a more complex People data file with different edge cases (ex: Use of apostrophes in a name, empty lists, alternative names, etc.)
Adding in Movie data with a new mapping file. Both of which expand the People data type and mapping. The People objects are now actors who are cast members of movies.