Update movies_schema_map.json
removed Genre as a node
Made Genre and Cast as properties of Movies
Also, kept cast as a edge linking movies to People. 

We may not want to keep cast as properties and as an edge. I kept it this way as there are way more cast then there are people in this small test data set.